fix mini-cart product display and cart count issue
